Update lowest EM mode to EM3 per spec Round float temperature to int

+++ src/i2c.h
@@ -22,7 +22,7 @@
 void initialize_I2C();
 
 // SI7021 functions
-float SI7021_get_temperature();
+int SI7021_get_temperature();
 float SI7021_convert_temp(uint16_t temp_code);
 
 void SI7021_start_measure_temp();
